aapi/sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aapi-sdk"
version = "0.1.0"
description = "Python SDK for AAPI (Agentic Action Protocol Interface)"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
authors = [
    { name = "AAPI Contributors", email = "hello@aapi.io" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
crypto = [
    "pynacl>=1.5.0",
    "canonicaljson>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.3.0",
    "mypy>=1.0.0",
]
langchain = [
    "langchain>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["aapi"]
